requirements.txt
streamlit
Pillow
pytesseract
speechrecognition
python-dotenv
google-generativeai
langchain
langchain-community
langchain-huggingface
pydantic
faiss-cpu
PyMuPDF
googletrans